Restore old ramdisk
I set up a script to auto save the old ramdisk then reenable it in Case of a fuckup with older devices boot looping. The script will need to be restarted after, but instead of a brick, you now have a brick that can be fixed in 10 minutes. I’m not sure if the setup app was renamed to a .txt or a .save file to make it invalid, so if it doesn’t work, change line 193 from setupapp.save to setupapp.txt.